kramdown: Optionally fall back to rouge
Due to the distributed development of tree-sitter parsers, it's not
always feasible to have a particular tree-sitter parser installed. In
some cases, we may choose to fall back to rouge if a tree-sitter parser
is not installed. This option defaults to false, but can be set via
syntax_highlighter_opts.